0.2.10 Textures loading without repetition
There is a small but: if you try to drag and drop from code, it won't work. There are comments to habilitate it in the LocalizePath function. It works if you do a build and put it in the folder that serves as base directory though: since the user won't have any problem, I won't fix it